+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- added: (LI.FI) Add SUI blockchain support for swaps via Aftermath and Bluefin7k DEXs
+
 ## 2.33.0 (2025-09-09)
 
 - added: Add Bridgeless plugin